backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
ftfy==6.1.1
regex==2023.10.3
tqdm==4.66.1
numpy==1.24.3
Pillow==10.0.0
faiss-cpu==1.7.4
python-dotenv==1.0.0
pydantic==2.4.2 